fix : 공연 엘라스틱서치
