+++ src/main/java/com/example/stagemate/service/performance/PerformanceCrawlingBatchService.java
@@ -41,6 +41,11 @@ public class PerformanceCrawlingBatchService {
 
         for (Performance performance : ongoingPerformances) {
             performance.updateStatusBasedOnCurrentDate();
+
+            if(performance.getPerformanceStatus() == PerformanceStatus.ENDED) {
+                searchService.deleteFromPerformanceId(performance.getId());
+            }
+
             performanceRepository.save(performance);
         }
     }
@@ -50,8 +55,6 @@ public class PerformanceCrawlingBatchService {
     @Transactional
     public void updateBatch(List<CrawledPerformanceInfo> crawledPerformances) {
         // 상영중, 상영예정 공연 가져오기
-
-        searchService.deleteAllFromPerformances();
 
         List<Performance> existingPerformances = performanceRepository.findByPerformanceStatusIn(
                 List.of(PerformanceStatus.ONGOING, PerformanceStatus.UPCOMING));
